+++ packages/esm-patient-attachments-app/src/attachments/delete-attachment.modal.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { Button, ModalHeader, ModalBody, ModalFooter } from '@carbon/react';
 import { type Attachment } from '@openmrs/esm-framework';
-import styles from './delete-attachment-confirmation-modal.scss';
+import styles from './delete-attachment.scss';
 
 interface DeleteAttachmentConfirmationProps {
   attachment: Attachment;
