+++ packages/esm-patient-attachments-app/src/attachments/delete-attachment.scss
