+++ packages/esm-patient-banner-app/src/banner-tags/print-identifier-sticker.scss
