+++ packages/esm-patient-chart-app/src/visit/visits-widget/past-visits-components/delete-encounter.modal.tsx
