+++ packages/esm-patient-conditions-app/src/conditions/delete-condition.modal.tsx
