+++ packages/esm-patient-vitals-app/src/vitals-biometrics-form/delete-vitals-and-biometrics.modal.tsx
