+++ packages/esm-patient-vitals-app/translations/en.json
@@ -60,7 +60,7 @@
   "Vitals & Biometrics": "Vitals & Biometrics",
   "vitalsAndBiometrics": "Vitals and biometrics",
   "vitalsAndBiometricsDeleted": "Vitals and Biometrics deleted",
-  "vitalsAndBiometricsDeleteError": "Error deleting vitals and biometrics",
+  "vitalsAndBiometricsDeleteError": "Error deleting Vitals and Biometrics",
   "vitalsAndBiometricsEditError": "Error editing vitals and biometrics",
   "vitalsAndBiometricsNowAvailable": "They are now visible on the Vitals and Biometrics page",
   "vitalsAndBiometricsRecorded": "Vitals and Biometrics saved",
